* Convert old socket/connection code to use smart pointers. * Prevent multiple connection requests from creating new sockets each time and blocking a client port.

+++ Engine/Core/Networking/ServerNetworkingManager.h
@@ -297,13 +297,12 @@ namespace networking
 			if (onNetworkViewDisconnectCallback != nullptr)
 				onNetworkViewDisconnectCallback(id);
 
-
 			if (disconnect)
 			{
 				std::shared_ptr<Address> owner = nullptr;
 				for (auto& kvp : addressToNetworkIDs)
 				{
-					for (int i = 0; i < kvp.second.size(); i++)
+					for (size_t i = 0; i < kvp.second.size(); i++)
 					{
 						if (kvp.second[i].lock()->GetUniqueID() == id)
 						{
